+++ packages/@withstudiocms/internal_helpers/src/utils/jsonUtils.ts
@@ -17,10 +17,7 @@ export function jsonParse<T extends object>(text: string): T {
  */
 export function readJson<T extends object>(
 	path: string | URL,
-	readFileSync: (
-		path: fs.PathOrFileDescriptor,
-		encoding: BufferEncoding
-	) => string = fs.readFileSync
+	readFileSync: (path: string | URL, encoding: BufferEncoding) => string = fs.readFileSync
 ): T {
 	const content = readFileSync(path, 'utf-8');
 	return jsonParse<T>(content);
